Creating Code for Maximum and Minimum
Created Python code to create variables "largest" and "smallest" and created a loop. Code asks for user input numbers and tries to convert these user input strings into an integer or gives an error.

If the number(s) are integers, it compares the newly input numbers to the “smallest” and “largest” variable- if it is less than “smallest”, the code sets "smallest" to new input and if it is larger than “largest”, the code sets "largest" to new input. When the user inputs “done”, the loop breaks and the program prints the "largest" and "smallest" numbers as the maximum and minimum.